Muutokset int -> float arvoiksi.

+++ Ohjelmisto-1-PModuulit/Python - Moduuli 2 tehtävät/Tehtävä 2.py
@@ -1,6 +1,6 @@
 #tuo matematiikan kirjaston
 import math
 #pyytää käyttäjältä numeron
-sade = int(input("Anna ympyrän säde. \n"))
+sade = float(input("Anna ympyrän säde. \n"))
 #printtaa säteen
 print(math.pi * sade**2)
